コマンド解析処理のファイル分割
